Begin refactor of email
