+++ mcmonarch_bot/src/lib.rs
@@ -1,6 +1,4 @@
 use color_eyre::eyre;
-use regex::Regex;
-use serde::{Deserialize, Serialize};
 use serenity::{
     async_trait,
     http::client,
@@ -11,74 +9,29 @@ use serenity::{
 };
 use std::env;
 
-#[derive(Debug, Serialize, Deserialize)]
-struct Email<'a> {
-    from: &'static str,
-    to: &'a str,
-    subject: &'static str,
-    template: &'static str,
-    #[serde(rename = "h:X-Mailgun-Variables")]
-    variables: String,
-}
-
-impl<'a> Email<'a> {
-    pub fn new(to: &'a str, vars: EmailVariables) -> eyre::Result<Self> {
-        let vars_ser = serde_json::to_string(&vars).map_err(|e| eyre::eyre!(e))?;
-
-        Ok(Self {
-            from: "CS @ ODU Discord <postmaster@mg.odu-cs-community.codes>",
-            to,
-            subject: "Verify your email",
-            template: "odu-cs-monarch-verify",
-            variables: vars_ser,
-        })
-    }
-
-    pub async fn send(&self) -> eyre::Result<()> {
-        let api_key = env::var("MCMONARCH_MAIL_API_KEY")?;
-
-        reqwest::Client::new()
-            .post("https://api.mailgun.net/v3/mg.odu-cs-community.codes/messages")
-            .basic_auth("api", Some(api_key))
-            .form(self)
-            .send()
-            .await
-            .map(|_| ())
-            .map_err(|e| eyre::eyre!(e))
-    }
-
-    // user portion of regular expression courtesy of https://html.spec.whatwg.org/multipage/input.html#valid-e-mail-address
-    fn extract_odu_email(msg: &str) -> Option<String> {
-        let re = Regex::new(r"[a-zA-Z0-9.!#$%&'*+/=?^_`{|}~-]+@odu\.edu").unwrap();
-        re.captures_iter(msg)
-            .next()
-            .map(|cap| cap.get(0).map(|cap| String::from(cap.as_str())))
-            .flatten()
-    }
-}
-
-#[derive(Debug, Serialize, Deserialize)]
-struct EmailVariables {
-    uid: Vec<u8>,
-}
-
-impl EmailVariables {
-    pub fn new(data: &str) -> eyre::Result<Self> {
-        let passphrase = env::var("MCMONARCH_CIPHER_PASSPHRASE")?;
-        let ciphered_uid = simplecrypt::encrypt(data.as_bytes(), passphrase.as_bytes());
-
-        Ok(Self { uid: ciphered_uid })
-    }
-}
-
-struct Handler;
+mod mailer;
+use mailer::{Email, EmailVariables};
 
 #[async_trait]
-impl EventHandler for Handler {
+impl EventHandler for McmonarchBot {
     async fn message(&self, ctx: Context, msg: Message) {
         if McmonarchBot::should_ignore(&ctx, &msg).await {
             return;
         }
+
+        /*
+        What I want this to look like
+        let email = Email::builder()
+            .to(Email::extract_odu_email(&msg.content))
+            .with_var("uid", &msg.author.id.to_string())
+            .build();
+        
+        match email {
+            Ok(email) => ...,
+            Err(InvalidEmailAddr) => ...,
+            Err(_) => ...,
+        }
+        */
 
         match Email::extract_odu_email(&msg.content) {
             Some(email_addr) => {
@@ -89,30 +42,24 @@ impl EventHandler for Handler {
                 let email = Email::new(&email_addr, vars).unwrap();
 
                 match email.send().await {
+                    Ok(res) if res.status() == reqwest::StatusCode::OK => {
+                        let _ = msg
+                        .reply(
+                            &ctx,
+                            ":ok_hand: \
+                        Check your email. I'm sending a verification link to that address. \n \
+                        (check your spam if it doesn't show up)",
+                        )
+                        .await;
+                    },
+                    Ok(res) => {
+                        self.reply_on_email_problem(&ctx, msg).await;
+                        log::warn!("{}", res.status());
+                    }
                     // This isn't doing what we want. we need to check status codes
                     Err(e) => {
-                        let reply = MessageBuilder::new()
-                            .push_line(":fire_engine: Uh-oh \n \
-                                Something went wrong sending your verification email. \n \
-                                Try again later maybe? If this keeps happening file an issue on gitlab.")
-                            .push_named_link(
-                                "CS @ ODU Meta gitlab", 
-                                "https://git-community.cs.odu.edu/community-discord/meta/-/issues/new"
-                            )
-                            .build();
-
-                        let _ = msg.reply(&ctx, reply).await;
+                        self.reply_on_email_problem(&ctx, msg).await;
                         log::warn!("{}", e);
-                    }
-                    _ => {
-                        let _ = msg
-                            .reply(
-                                &ctx,
-                                ":ok_hand: \
-                            Check your email. I'm sending a verification link to that address. \n \
-                            (check your spam if it doesn't show up)",
-                            )
-                            .await;
                     }
                 };
             }
@@ -170,7 +117,7 @@ impl EventHandler for Handler {
     }
 }
 
-pub struct McmonarchBot {}
+pub struct McmonarchBot;
 
 impl McmonarchBot {
     pub async fn verify(data: Vec<u8>) -> eyre::Result<()> {
@@ -196,71 +143,27 @@ impl McmonarchBot {
     pub async fn should_ignore(cache: impl AsRef<serenity::cache::Cache>, msg: &Message) -> bool {
         !msg.is_private() || msg.is_own(cache).await
     }
+
+    async fn reply_on_email_problem(&self, ctx: &Context, msg: Message) {
+        let reply = MessageBuilder::new()
+            .push_line(":fire_engine: Uh-oh \n \
+                Something went wrong sending your verification email. \n \
+                Try again later maybe? If this keeps happening file an issue on gitlab.")
+            .push_named_link(
+                "CS @ ODU Meta gitlab", 
+                "https://git-community.cs.odu.edu/community-discord/meta/-/issues/new"
+            )
+            .build();
+
+        let _ = msg.reply(&ctx, reply).await;
+    }
 }
 
 pub async fn get_bot(token: &str) -> eyre::Result<()> {
     let mut client = Client::new(token)
-        .event_handler(Handler)
+        .event_handler(McmonarchBot)
         .await
         .expect("Err creating client");
 
     client.start().await.map(|_| ()).map_err(|e| eyre::eyre!(e))
-}
-
-#[cfg(test)]
-mod test {
-    use super::*;
-
-    #[test]
-    fn extracts_odu_valid_email() {
-        let expected = Some("monarchsrule@odu.edu".to_string());
-
-        assert_eq!(expected, Email::extract_odu_email("monarchsrule@odu.edu"));
-        assert_eq!(
-            expected,
-            Email::extract_odu_email("my email is monarchsrule@odu.edu")
-        );
-        assert_eq!(
-            expected,
-            Email::extract_odu_email("monarchsrule@odu.edu. hmu!! 💝")
-        );
-        assert_eq!(
-            expected,
-            Email::extract_odu_email("my email is monarchsrule@odu.edu cash me outside")
-        );
-    }
-
-    #[test]
-    fn fails_invalid_odu_email() {
-        assert_eq!(None, Email::extract_odu_email(""));
-        assert_eq!(None, Email::extract_odu_email("i_love_guac@hotmail.com"));
-    }
-
-    #[test]
-    #[ignore]
-    fn send_email() -> eyre::Result<()> {
-        let api_key = dotenv::var("MCMONARCH_MAIL_API_KEY")?;
-        std::env::set_var("MCMONARCH_MAIL_API_KEY", api_key);
-        let email = Email::new(
-            "alex.launi@gail.com",
-            EmailVariables::new("dave grohl sucks")?,
-        )?;
-        tokio_test::block_on(email.send())?;
-
-        std::env::set_var("MAIL_API_KEY", "");
-        Ok(())
-    }
-
-    #[test]
-    fn test_encrypt_decrypt() {
-        let key = dotenv::var("MCMONARCH_CIPHER_PASSPHRASE").unwrap();
-        std::env::set_var("MCMONARCH_CIPHER_PASSPHRASE", &key);
-        let plaintext = "alaun001@odu.edu".as_bytes();
-        let cipher = simplecrypt::encrypt(plaintext, key.as_bytes());
-        assert_eq!(
-            plaintext,
-            simplecrypt::decrypt(&cipher, key.as_bytes()).unwrap()
-        );
-        std::env::set_var("MCMONARCH_CIPHER_PASSPHRASE", "");
-    }
 }
